backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# AI/ML
sentence-transformers==2.5.1
transformers==4.37.2
torch==2.1.2
numpy==1.26.3

# Resume Parsing
pdfplumber==0.10.3
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.2.0

# NLP
spacy==3.7.2
nltk==3.8.1

# Optional LLM
openai==1.10.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
